dockerized setup for sharing CI scripts between plugins (#1)
this PR is the initial work (extracted from existing plugins) to provide some level of CI sharing.

the idea is for this sub-repo to be checked out from a plugin and its (existing) ci/ folder should re-use as much as possible - simply by linking whole folders or individual file (git is expected to resolve sym-links even on Windows, if not e.g. due FAT its still configurable to use hard copies).