Pass stack version as argument to image fetch function
This script reads and modifies environment variables for the processes it
controls. This commit changes the way the `pull_docker_snapshot` function
references variables. Specifically, we now explicitly pass arguments that are
*not* managed as environment variables by the script as explicit parameters to
the script instead of looking them up from global scope. The pattern this
attempts to acheive is that the only thing looked up from global scope in a
function is an argument that is actively being managed by the environment such
that it remains consistent across any sub processes spawned by the shell script.