Update variable name to be more descriptive
This commit updates the script to be more descriptive with the naming of a
variable that indicates the stack version. The stack version may be an alias
when provided to this script. Using variable names to track when we are dealing
with a potential alias (vs when we have normalized) helps with maintainability.